Mid Section 2
I am making this commit early, due to not having worked on the project for
a few weeks. At this point the requirements for finishing Section 2 are:
completing the catacombs tileset, and the two remaining stages. Not much
work, just not enough motivation to work on it lately.

Who am I writing this for?